+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.6.3] - 2021-03-22
+
+### Added
+
+- Add `extension::OpenTelemetry`.
+
+### Removed
+
+- Remove `TracingConfig`, now Request span always takes the current span as the parent, so this option is no longer needed.
+- Remove `multipart` feature.
+
+### Changed
+
+- Now all features are not activated by default.
+
 ## [2.6.2] - 2021-03-20
 
 - Add `SchemaBuilder::enable_subscription_in_federation` method.  [#449](https://github.com/async-graphql/async-graphql/issues/449)